igc: Add BTF based metadata for XDP
This commit adds support for BTF based metadata for XDP. Currently, the
support has only been tested on receive side. Following is the struct
describing the metadata:

struct xdp_md_desc {
	u64 timestamp;
};

Note that only a single member is added to the struct. More members will
be added in the future.

Signed-off-by: Vedang Patel <vedang.patel@intel.com>